Fix license headers: Should have been AGPLv3+, not GPLv2+
I'm not sure why so many files (particularly written by Neels)
did contain a GPLv2+ header, instead of the AGPLv3+ which is the
actual overall project license.  I consider it a mistake.

In any case, any copyrightable contribution to those files was done by
sysmocom employees, so I as managing directory can legally make a
license change, whther or not it was a mistake early on or not.

The only GPLv2-or-later file remaining is mncc_internal.c, as it has
more contributors and a longer history.

Change-Id: I8650697592b3160c4d0a7c61ae9c46d4aacb3bef