local hack: unconditionally request ThemWi RTP extensions
This patch is NOT intended for merging into Osmocom mainline,
instead it is intended for experimentation/test use by interested
members of Osmocom community, and possibly for production use
at ThemWi.  Furthermore, despite residing on a branch that is
based on the elderly 2023-02 release version of OsmoMSC, this
patch requires libosmocore changes that were merged in mid-2024.
The intended build configuration is a combination of the present
branch version of osmo-msc with current Osmocom libraries.

Related: OS#6448
Depends: I0eccfe5ddcf44f8f20440acb01e2d4870ec0cd91 (libosmocore)
Change-Id: Ie10478724ea0870bc2b88bdbaa02f6c95c89b9d8