RFC: vlr: vlr_lu_fsm: update the HLR only on initial IMSI attach
As long the VLR isn't supporting multiple MSCs, skip updating the HLR
on PERIODIC and REGULAR Location Updates (same for PS Attach Request/Routing Area Update).

TODO: fix testcases as soon it is clear, if this should be fixed (or configurable).
The SGSN currently works by following the spec and requires this commit.

Change-Id: I7adc2092557f8cc6bb358234df2d1a16d1036eee