WIP: vlr: extend the subscriber invalidate callback with reasons
The VLR must be allowed to notify the MSC if an subscriber becomes invalid.
There are multiple cases when this happens:
a) if the subscriber didn't do a Location Update Procedure within the
given periodic timer.
b) if the HLR do a Cancel Location Procedure with reason withdraw
c) if the HLR do a Cancel Location Procedure with reason update location.

Only in case b) must the user informed with the cause code by the HLR.
For c) the MSC must inform the MS about the reason.

FIXME: check case a) what must be done. Do we really need to inform the user?
Change-Id: Ie5b687318b106a230fcee52deba86649641004b3