Release BSS connection when CC transaction fails
When the creation of the transaction structure fails, the MSC will
send a RELEASE COMPLETE message.

This patch ensures that the MSC_A use counter is decremented, so
that the BSS connection is released, if it is not used by any other
transaction.

Without this patch, the msc_a_fsm would wait 5 seconds for an initial
transaction before releasing the BSS connection.

Related: OS#6427
Change-Id: Icb893d5c1ce4c77a29259ce66189576dcaa5a44a