ran_peer: Add specific API to free object
Add a specific ran_peer_free() API so it becomes clear where the object
is freed and hence the entire lifecycle of the object.
Swap the talloc tree parentship to have the fi be a child of the object,
which simplifies tear down triggered by FSM cleanup.

Apparently the object is not being freed anywhere, so they are so far
only being added to the list through .ran_peer_up_l2() -> ran_peer_find_or_create().

Change-Id: I721de21a68a4e336ae508a995e3cfcca05d57efe