sqlite optimisation: check VLR earlier for dest. subscriber
sms_from_result() calls a number of sqlite3_column_XXX functions
to build the SMS in memory.

If the destination subscriber is not attached, the sms
will be free'd immediatly anyway in smsq_take_next_sms()
so let's check if the destination subscriber is present
and attached before we call all the sqlite3 routines.

Change-Id: Ibd07b9d200b48108d705ed4c461cc4ddfa421fd2