ASCI: Fix establishment of calls when ciphering is enabled
We do not support any VGCS/VBS ciphering. If the CM Service Request
indicates VGCS or VBS service, encryption is not enabled for the channel
to be set up. There will be no Ciphering Command sent to the ME.

Authentication is still performed with the initiator of the call, if
authentication is enabled. Only authenticated subscribers will then be
able to initiate ASCI calls.

Change-Id: Ic59221e6b0f64c9f19508ae31cc3672548d0666e